More-cloud-test-coverage (#2229)
* SQC tests

* Raise unsupported op when ack on SQC

* Return 0, 0, 0 for SQC version

* Add admin user

* Configurable admin user

* Formatting

* Fix test API for SQC

* Add org for SQC

* Add org

* Add default groups

* Revert to API v1 for users

* Raise unsupported op for not existing API

* Return None to create groups and users

* SQC admin group

* Fix for SQC

* Admin group

* Admin group

* Admin group

* Admin group

* Remove user, group, portfolios and app tests for SQC

* Fix audit pattern for SQC

* SQC compat

* SQC compat

* Cleanup reload of tasks()

* Fix reload()

* Fix reload

* Add create delete conditions

* Pass id for delete

* Collect id for delete on SQC

* Fixes #2227

* Remove unneeded code

* use return _field from JSON

* Make count a classmethod

* Delegate count in sqobject

* use PROJECT_FILTER variable

* Adjust to new count

* Adjust to new count

* Inherit issue count()

* Implement search_one_page() for all objects supporting load()

* Move load_objects() as a classmethod

* Formatting

* Formatting

* Type hint

* Fix test

* Fix AI regression :-(

* Fix regression

---------

+++ test/unit/conftest.py
@@ -215,20 +215,27 @@ def get_test_user() -> Generator[users.User]:
     try:
         o = users.User.get_object(endpoint=tutil.SQ, login=key)
     except exceptions.ObjectNotFound:
-        o = users.User.create(endpoint=tutil.SQ, login=key, name=f"User name {key}")
-    (uid, uname, ulogin) = (o.name, o.id, o.login)
-    for g in o.groups():
-        if g != tutil.SQ.default_user_group():
-            o.remove_from_group(g)
-    yield o
-    try:
-        (o.name, o.id, o.login) = (uid, uname, ulogin)
+        try:
+            o = users.User.create(endpoint=tutil.SQ, login=key, name=f"User name {key}")
+        except exceptions.UnsupportedOperation:
+            if not tutil.SQ.is_sonarcloud():
+                raise
+            o = None
+    if o:
+        (uid, uname, ulogin) = (o.name, o.id, o.login)
         for g in o.groups():
             if g != tutil.SQ.default_user_group():
                 o.remove_from_group(g)
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    yield o
+    if o:
+        try:
+            (o.name, o.id, o.login) = (uid, uname, ulogin)
+            for g in o.groups():
+                if g != tutil.SQ.default_user_group():
+                    o.remove_from_group(g)
+            o.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 def rm(file: str) -> None:
@@ -311,24 +318,31 @@ def get_test_group() -> Generator[groups.Group]:
     try:
         o = groups.Group.get_object(endpoint=tutil.SQ, name=tutil.TEMP_KEY)
     except exceptions.ObjectNotFound:
-        o = groups.Group.create(endpoint=tutil.SQ, name=tutil.TEMP_KEY)
+        try:
+            o = groups.Group.create(endpoint=tutil.SQ, name=tutil.TEMP_KEY)
+        except exceptions.UnsupportedOperation:
+            if not tutil.SQ.is_sonarcloud():
+                raise
+            o = None
     yield o
-    try:
-        o.delete()
-    except exceptions.ObjectNotFound:
-        pass
+    if o:
+        try:
+            o.delete()
+        except exceptions.ObjectNotFound:
+            pass
 
 
 @pytest.fixture
 def get_60_groups() -> Generator[list[groups.Group]]:
     group_list = []
-    for i in range(60):
-        gr_name = f"Group-{tutil.TEMP_KEY}{i}"
-        try:
-            o_gr = groups.Group.get_object(endpoint=tutil.SQ, name=gr_name)
-        except exceptions.ObjectNotFound:
-            o_gr = groups.Group.create(endpoint=tutil.SQ, name=gr_name, description=gr_name)
-        group_list.append(o_gr)
+    if not tutil.SQ.is_sonarcloud():
+        for i in range(60):
+            gr_name = f"Group-{tutil.TEMP_KEY}{i}"
+            try:
+                o_gr = groups.Group.get_object(endpoint=tutil.SQ, name=gr_name)
+            except exceptions.ObjectNotFound:
+                o_gr = groups.Group.create(endpoint=tutil.SQ, name=gr_name, description=gr_name)
+            group_list.append(o_gr)
     yield group_list
     for g in group_list:
         g.delete()
